refactor: awsconfig package
