+++ internal/config/locations/awsconfig/get-path.go
@@ -1,4 +1,4 @@
-package locations
+package awsconfig
 
 import (
 	"os"
@@ -8,7 +8,7 @@ import (
 // Get AWS config file which usually is under .aws folder in user home directory,
 // unless user has set $AWS_CONFIG_FILE environment variable.
 // https://docs.aws.amazon.com/cli/latest/userguide/cli-configure-files.html#cli-configure-files-where
-func getAwsConfigFile() (string, error) {
+func GetPath() (string, error) {
 	if val := os.Getenv("AWS_CONFIG_FILE"); val != "" {
 		return val, nil
 	}
