+++ internal/config/locations/awsconfig/get-path_test.go
@@ -1,4 +1,4 @@
-package locations
+package awsconfig
 
 import (
 	"os"
@@ -8,10 +8,10 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestLocatioAwsConfigDir(t *testing.T) {
+func TestGetPath(t *testing.T) {
 
 	t.Run("default unix", func(t *testing.T) {
-		actual, err := getAwsConfigFile()
+		actual, err := GetPath()
 		homeDir, _ := os.UserHomeDir()
 		assert.Equal(t, err, nil)
 		assert.Equal(t, filepath.Join(homeDir, ".aws", "config"), actual)
@@ -20,7 +20,7 @@ func TestLocatioAwsConfigDir(t *testing.T) {
 	t.Run("$AWS_CONFIG_FILE set", func(t *testing.T) {
 		customLocation := "/tmp/custom/config"
 		os.Setenv("AWS_CONFIG_FILE", customLocation)
-		actual, err := getAwsConfigFile()
+		actual, err := GetPath()
 		assert.Equal(t, err, nil)
 		assert.Equal(t, customLocation, actual)
 	})
